Update nav and fixed initial load
