+++ package.json
@@ -14,6 +14,7 @@
         "lint": "eslint src --ext '.js,.jsx' --config .eslintrc",
         "lint:fix": "npm run lint --fix",
         "deploy:staging": "npm run build && aws s3 sync build/ s3://entaxy-staging",
+        "deploy:production": "npm run build && aws s3 sync build/ s3://entaxy-production",
         "deploy:test": "npm run build && aws s3 sync build/ s3://entaxy-test"
     },
     "dependencies": {
